requirements.txt
fastapi
uvicorn
clipspy
httpx
python-dotenv
